Break Layout component into smaller components
