--- gatsby/src/components/Layout.module.scss
@@ -1,152 +1,0 @@
-@import '../../styles/variables';
-@import '../../styles/global';
-
-$sidebar-width: 260px;
-$transition: all 0.15s ease-out;
-
-.header {
-  background: $mm-background;
-  border-bottom: 1px solid $mm-border;
-  height: var(--page-header-height);
-  position: sticky;
-  top: 0;
-  width: 100%;
-  z-index: 10;
-
-  &::before {
-    content: ' ';
-    display: block;
-    height: 10px;
-    left: 0;
-    overflow: hidden;
-    padding: 0 !important;
-    position: absolute;
-    top: 0;
-    width: 100%;
-
-    @include pattern;
-  }
-
-  &::after {
-    background: rgb(0, 119, 194);
-    background: linear-gradient(90deg, rgba(0, 119, 194, 0.5) 0%, var(--page-type-color));
-    content: ' ';
-    display: block;
-    height: 10px;
-    left: 0;
-    overflow: hidden;
-    padding: 0 !important;
-    position: absolute;
-    top: 0;
-    width: 100%;
-  }
-}
-
-.toggle {
-  background: $mm-background;
-  border: 1px solid $mm-border;
-  border-radius: 0;
-  color: #333;
-  height: 40px;
-  left: 0;
-  position: absolute;
-  top: 10px;
-  width: 50px;
-  z-index: 4;
-}
-
-.main {
-  background: $mm-background;
-  display: grid;
-  grid-template-areas: 'sidebar article';
-  grid-template-columns: $sidebar-width calc(100vw - #{$sidebar-width});
-  transition: $transition;
-}
-
-.sidebar {
-  background: #f5f6f7;
-  border-right: 1px solid $mm-border;
-  grid-area: sidebar;
-  position: relative;
-  transition: $transition;
-  z-index: 9;
-}
-
-.sidebarNav {
-  position: sticky;
-  top: calc(var(--page-header-height) + #{$mm-spacing});
-}
-
-.sidebarNavItem {
-  border-bottom: 1px solid $mm-border;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
-  margin-bottom: $mm-spacing / 4;
-}
-
-.sidebarNavLink {
-  background: #fff;
-  border-top: 1px solid #eee;
-  color: #666;
-  display: inline-block;
-  font-family: $mm-font-display;
-  font-size: 14px;
-  font-weight: 500;
-  height: 60px;
-  line-height: 60px;
-  padding: 0 $mm-spacing / 2;
-  text-decoration: none;
-  transition: all 0.25s ease-out;
-  width: 100%;
-
-  &:hover {
-    background: rgba(255, 255, 255, 0.5);
-    padding: 0 $mm-spacing / 2 0 $mm-spacing;
-  }
-}
-
-.sidebar--hidden .sidebar {
-  transform: translate3d(-100%, 0, 0);
-}
-
-.content {
-  grid-area: article-start / sidebar-start / article-end / article-end;
-  transition: $transition;
-}
-
-.footer {
-  border-top: 1px solid $mm-border;
-  font-size: 14px;
-  padding: $mm-spacing 0;
-  text-align: center;
-}
-
-@include breakpoint('lg') {
-  $sidebar-width: 350px;
-
-  .main {
-    grid-template-columns: $sidebar-width calc(100vw - #{$sidebar-width});
-  }
-
-  .toggle {
-    display: none;
-  }
-
-  .sidebar {
-    grid-area: sidebar;
-    transform: initial !important;
-    transition: none;
-  }
-
-  .sidebarNavLink {
-    font-size: 16px;
-    padding: 0 $mm-spacing;
-
-    &:hover {
-      padding-left: $mm-spacing * 1.25;
-    }
-  }
-
-  .content {
-    grid-area: article;
-  }
-}
