+++ gatsby/src/components/Layout/Footer.module.scss
@@ -1,0 +1,8 @@
+@import '../../styles/variables';
+
+.footer {
+  border-top: 1px solid $mm-border;
+  font-size: 14px;
+  padding: $mm-spacing 0;
+  text-align: center;
+}
