+++ gatsby/src/components/Layout/Header.module.scss
@@ -1,0 +1,58 @@
+@import '../../styles/variables';
+
+.header {
+  background: $mm-background;
+  border-bottom: 1px solid $mm-border;
+  height: var(--page-header-height);
+  position: sticky;
+  top: 0;
+  width: 100%;
+  z-index: 10;
+
+  &::before {
+    content: ' ';
+    display: block;
+    height: 10px;
+    left: 0;
+    overflow: hidden;
+    padding: 0 !important;
+    position: absolute;
+    top: 0;
+    width: 100%;
+
+    @include pattern;
+  }
+
+  &::after {
+    background: rgb(0, 119, 194);
+    background: linear-gradient(90deg, rgba(0, 119, 194, 0.5) 0%, var(--page-type-color));
+    content: ' ';
+    display: block;
+    height: 10px;
+    left: 0;
+    overflow: hidden;
+    padding: 0 !important;
+    position: absolute;
+    top: 0;
+    width: 100%;
+  }
+}
+
+.toggle {
+  background: $mm-background;
+  border: 1px solid $mm-border;
+  border-radius: 0;
+  color: #333;
+  height: 40px;
+  left: 0;
+  position: absolute;
+  top: 10px;
+  width: 50px;
+  z-index: 4;
+}
+
+@include breakpoint('lg') {
+  .toggle {
+    display: none;
+  }
+}
