+++ gatsby/src/pages/404.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Layout from '../components/Layout';
+import Layout from '../components/Layout/Layout';
 
 const NotFoundPage: React.FC = () => (
   <Layout
