+++ gatsby/src/pages/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Layout from '../components/Layout';
+import Layout from '../components/Layout/Layout';
 
 const IndexPage: React.FC = () => (
   <Layout
